API DEVELOPMENT OF SWEETSPOT/FrontEnd/requirements.txt
streamlit>=1.10.0
streamlit-extras>=0.3.0
requests
Pillow
psycopg2-binary